implementação do AND + correções de warnings g++

+++ src/Memory.cpp
@@ -30,7 +30,7 @@ void Memory::fillRandomData() {
 }
 
 void Memory::fillSequencialData() {
-  for (int i = 0; i < data.size(); i++) {
+  for (size_t i = 0; i < data.size(); i++) {
     data[i] = i;
   }
 }
